requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
beautifulsoup4>=4.12.2
lxml>=5.2.2
tenacity>=8.2.3
python-dateutil>=2.9.0
jinja2>=3.1.3
openpyxl>=3.1.2
streamlit>=1.44.0
pandas>=2.2.3
